Support inline sources
The package now can handle situations where the source files are located in
the source map.  The main, grunt and gulp automatically handle this.  While
the remap module can take a Istanbul memory store which is then used to
populate the sources to then be passed to reports the require the source
files (like the Istanbul HTML report).

Fixes #11